aider: Update the DATE variable to store the current date and time in ISO 8601 format.
# Context:
USER: update the DATE variable to store the current date and time in iso8601 format
ASSISTANT: To update the DATE variable to store the current date and time in ISO 8601 format, you need to modify the makefile. Here is the change:

```makefile
makefile
<<<<<<< ORIGINAL
DATE := $(shell date)
=======
DATE := $(shell date -u +"%Y-%m-%dT%H:%M:%SZ")
>>>>>>> UPDATED
```

This change will update the date command to output the date and time in ISO 8601 format. The "-u" option is used to get the date and time in UTC. The "+%Y-%m-%dT%H:%M:%SZ" option formats the date and time in the desired format.