aider: Update the Test step to send the `go test` command output to the screen and also to the `test-report.out` file.
# Aider chat conversation:

USER: update the step named Test to send the `go test` command output to the screen and also to the `test-report.out` file.

ASSISTANT: To achieve this, we need to modify the `.github\workflows\build.yml` file. Specifically, we need to update the `run` command in the `Test` step under the `test` job to use the `tee` command. The `tee` command will allow us to send the output of the `go test` command to both the console and the `test-report.out` file.

Here is the change:

```yaml
.github\workflows\build.yml
<<<<<<< ORIGINAL
      - name: Test
        run: go test -v -race -coverprofile=coverage.out -covermode=atomic -json ./... > test-report.out
=======
      - name: Test
        run: go test -v -race -coverprofile=coverage.out -covermode=atomic -json ./... | tee test-report.out
>>>>>>> UPDATED
```

This change will make the output of the `go test` command visible in the console during the GitHub Actions run, and it will also be saved in the `test-report.out` file.